Moved post loads on /Blog access
