+++ src/App.vue
@@ -1,61 +1,9 @@
 <script setup>
 import { RouterView } from "vue-router"
 import { isDarkMode } from './utils/LightSwitch'
-import { ref, onMounted } from 'vue';
-import { updateBlogs } from './plugins/store'
-import { gql } from "graphql-request";
+import { onMounted } from 'vue';
 
-import { request } from './utils/datocms';
-
-const blogPosts = ref([])
-const error = ref(null)
-const loading = ref(true)
-
-const query = gql`
-{
-  allBlogs {
-    id
-    title
-    posted
-    blurb
-    content {
-      value
-      blocks {
-        __typename
-        ... on ImageRecord {
-          id
-          image {
-            responsiveImage {
-              src
-              width
-              height
-              alt
-              title
-              base64
-            }
-          }
-        }
-      }
-    }
-    preview
-  }
-}
-`
-
-onMounted(async () => {
-  try {
-    blogPosts.value = await request({
-      query,
-      variables: {
-        limit: 10
-      }
-    });
-    updateBlogs(blogPosts)
-  } catch (e) {
-    error.value = e;
-  }
-  loading.value = false;
-
+onMounted(() => {
   if (localStorage.theme === 'dark' || (!('theme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches)) {
     isDarkMode.value = true;
     document.body.style.backgroundColor = '#1a1a1a'
@@ -64,14 +12,10 @@ onMounted(async () => {
     document.body.style.backgroundColor = '#ede7de'
   }
 })
-
-
 </script>
 
 <template>
   <div :class="isDarkMode ? 'dark' : ''">
-    <div v-if="loading">Loading...</div>
-    <div v-else-if="error">Error</div>
-    <RouterView v-else class='bg-light text-dark dark:bg-dark dark:text-light transition-colors duration-200'/>
+    <RouterView class='bg-light text-dark dark:bg-dark dark:text-light transition-colors duration-200'/>
   </div>
 </template>
